+++ handler_testing_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestTestHandler_Handle(t *testing.T) {
+func TestTestHandler_Count(t *testing.T) {
 	handler := NewTestHandler()
 
 	err1 := errors.New("error 1")
@@ -17,13 +17,10 @@ func TestTestHandler_Handle(t *testing.T) {
 	handler.Handle(err1)
 	handler.Handle(err2)
 
-	errors := handler.Errors()
-
-	assert.Equal(t, err1, errors[0])
-	assert.Equal(t, err2, errors[1])
+	assert.Equal(t, 2, handler.Count())
 }
 
-func TestTestHandler_Last(t *testing.T) {
+func TestTestHandler_LastError(t *testing.T) {
 	handler := NewTestHandler()
 
 	err1 := errors.New("error 1")
@@ -32,11 +29,20 @@ func TestTestHandler_Last(t *testing.T) {
 	handler.Handle(err1)
 	handler.Handle(err2)
 
-	assert.Equal(t, err2, handler.Last())
+	assert.Equal(t, err2, handler.LastError())
 }
 
-func TestTestHandler_Last_Empty(t *testing.T) {
+func TestTestHandler_Errors(t *testing.T) {
 	handler := NewTestHandler()
 
-	assert.NoError(t, handler.Last())
+	err1 := errors.New("error 1")
+	err2 := errors.New("error 2")
+
+	handler.Handle(err1)
+	handler.Handle(err2)
+
+	errs := handler.Errors()
+
+	assert.Equal(t, err1, errs[0])
+	assert.Equal(t, err2, errs[1])
 }