+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Refactored and renamed nopHandler to noopHandler
+- Refactored the test handler
 
 
 ## [0.15.0] - 2018-12-22